move gameStatus to game state
