+++ browser/reducers/gameState.js
@@ -1,7 +1,7 @@
 /*-------- ACTION TYPES  ---------*/
 
 //const IS_DISPLAYED = 'IS_DISPLAYED';
-const START_GAME ='START_GAME';
+const START_GAME = 'START_GAME';
 const STOP_GAME = 'STOP_GAME';
 const START_CHAT = 'START_CHAT';
 const STOP_CHAT = 'STOP_CHAT';
@@ -9,7 +9,10 @@ const SET_ERROR = 'SET_ERROR';
 const SET_NICKNAME = 'SET_NICKNAME';
 const RESET_ERROR = 'RESET_ERROR';
 const HIDE_INSTRUCTIONS = 'HIDE_INSTRUCTIONS';
-
+const LOSE = 'LOSE';
+const CONTINUE = 'CONTINUE';
+const ATE = 'ATE';
+const FELL = 'FELL';
 
 /*=------ACTION CREATORS-------*/
 
@@ -17,6 +20,7 @@ const initialState = {
   isPlaying: false,
   isChatting: false,
   nickname: '',
+  status: '',
   error: null
 };
 
@@ -59,6 +63,27 @@ export const hideInstructions = () => ({
   type: HIDE_INSTRUCTIONS
 });
 
+export const lose = eater => {
+  return ({
+    type: LOSE,
+    eater: eater.length > 15 ? eater.slice(0, 14) + '...' : eater
+  });
+};
+
+export const keepPlaying = () => ({
+  type: CONTINUE
+});
+export const fell = world => ({
+  type: FELL,
+  eaten: world
+});
+export const ateSomeone = eaten => {
+  return ({
+    type: ATE,
+    eaten: eaten.length > 15 ? eaten.slice(0, 14) + '...' : eaten
+  });
+};
+
 /*----------  THUNK CREATORS  ----------*/
 
 export const startAsGuest = (nickname, socket) => dispatch => {
@@ -87,6 +112,7 @@ export default (state = initialState, action) => {
     case START_GAME:
       return Object.assign({}, state, {isPlaying: true});
     case STOP_GAME:
+      console.log('stopping game', action);
       return Object.assign({}, state, {isPlaying: false});
     case START_CHAT:
       return Object.assign({}, state, {isChatting: true});
@@ -99,8 +125,16 @@ export default (state = initialState, action) => {
     case SET_NICKNAME:
     return Object.assign({}, state, { nickname: action.text });
     case HIDE_INSTRUCTIONS:
-    return Object.assign({}, state, { instructionsHidden: true });
+      return Object.assign({}, state, { instructionsHidden: true });
+    case LOSE:
+      return Object.assign({}, state, { status: `${action.eater} rolled you up`});
+    case FELL:
+      return Object.assign({}, state, { status: `you left ${action.eaten}'s orbit`});
+    case CONTINUE:
+      return Object.assign({}, state, { status: ''});
+    case ATE:
+      return Object.assign({}, state, { status: `you rolled up ${action.eaten}`});
     default:
-    return state;
+      return state;
   }
 };
