+++ browser/reducers/index.js
@@ -3,7 +3,6 @@ import auth from './auth';
 import controlPanel from './controlPanel';
 import players from './players';
 import food from './food';
-import gameStatus from './gameStatus';
 import abilities from './abilities';
 import gameState from './gameState';
 import record from './record';
@@ -15,7 +14,6 @@ export default combineReducers({
   controlPanel,
   players,
   food,
-  gameStatus,
   abilities,
   gameState,
   record,
